+++ tests/Mockery/MockClassWithMethodOverloadingTest.php
@@ -3,6 +3,7 @@
 namespace test\Mockery;
 
 use Mockery\Adapter\Phpunit\MockeryTestCase;
+use Mockery\Exception\BadMethodCallException;
 
 class MockClassWithMethodOverloadingTest extends MockeryTestCase
 {
@@ -12,9 +13,17 @@ class MockClassWithMethodOverloadingTest extends MockeryTestCase
             ->makePartial();
         $this->assertInstanceOf('test\Mockery\TestWithMethodOverloading', $mock);
 
-        $this->expectException(\BadMethodCallException::class);
+        $this->assertSame(42, $mock->theAnswer());
+    }
 
-        // TestWithMethodOverloading::__call wouldn't be used. See Gotchas!.
+    public function testThrowsWhenMethodDoesNotExist()
+    {
+        $mock = mock('test\Mockery\TestWithMethodOverloadingWithoutCall')
+            ->makePartial();
+        $this->assertInstanceOf('test\Mockery\TestWithMethodOverloadingWithoutCall', $mock);
+
+        $this->expectException(BadMethodCallException::class);
+
         $mock->randomMethod();
     }
 
@@ -32,11 +41,15 @@ class TestWithMethodOverloading
 {
     public function __call($name, $arguments)
     {
-        return 1;
+        return 42;
     }
 
     public function thisIsRealMethod()
     {
         return 1;
     }
+}
+
+class TestWithMethodOverloadingWithoutCall
+{
 }