+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+* Allow passing thru __call method in all mock types (experimental) (#969)
+
 ## 1.2.2 (2019-02-13)
 
 * Fix a BC breaking change for PHP 5.6/PHPUnit 5.7.27 (#947)